Refactored to run agol_to_pg in flask instead of external process
Updated sql to handle errors vs failed to truncate because of foreign key.
Added vscode settings and launch to debug flask app inside vscode.
Update to use conn and cursor context for auto closing
Setup agol_to_pg to log to container logs as well as stream events
Added RESTART IDENTITY to sql in order to reset the main table sequence when truncated
Update OID to be required and not an existing field in the service to maintain OID values from the service
Moved API key validation to before request call so it can be auto re-used

+++ lib/sql.py
@@ -24,52 +24,66 @@ from psycopg2 import sql
 # Configure logging
 logging.basicConfig(level=logging.INFO)
 
-def truncate_or_delete_table(table_name, service_name, schema='public', cascade: bool = False):
+
+def truncate_or_delete_table(table_name, service_name, schema='public', cascade: bool = False, reset_sequence: bool = False):
     """
     Truncates the specified table in the PostgreSQL database using a service definition.
     If truncating fails (e.g., due to foreign key constraints), delete all rows instead.
-    
+
     Args:
     table_name (str): The name of the table to truncate or delete from.
     service_name (str): The PostgreSQL service name as defined in pg_service.conf.
     schema (str): The database schema in which the table resides. Default is 'public'.
     """
-    try:
-        # Establish a connection using the service name
-        conn = psycopg2.connect(f"service={service_name}")
+
+    # Establish a connection using the service name
+    with psycopg2.connect(f"service={service_name}") as conn:
         conn.autocommit = True  # Enable autocommit for DDL commands like TRUNCATE
-        cur = conn.cursor()
-
-        # Construct the TRUNCATE SQL query
-        query = sql.SQL("TRUNCATE TABLE {}.{}{}").format(
-            sql.Identifier(schema),
-            sql.Identifier(table_name),
-            sql.SQL(" CASCADE" if cascade else sql.SQL(""))
-        )
-
-        # Execute the truncate operation
-        cur.execute(query)
-        logging.info(f"Table {schema}.{table_name} truncated successfully.")
-
-    except Exception as e:
-        # If an error occurs during truncation, log the error and attempt to delete instead
-        logging.error(f"Failed to truncate table {schema}.{table_name}, perhaps because of a dependency within the database. Trying to delete features without truncating. Error: {str(e)}")
-        
         try:
-            # Construct the DELETE SQL query
-            delete_query = sql.SQL("DELETE FROM {}.{}").format(
-                sql.Identifier(schema),
-                sql.Identifier(table_name)
-            )
+            with conn.cursor() as cur:
+                # Construct the TRUNCATE SQL query
+                query = sql.SQL("TRUNCATE TABLE {}.{}{}{}").format(
+                    sql.Identifier(schema),
+                    sql.Identifier(table_name),
+                    sql.SQL(" RESTART IDENTITY" if reset_sequence else ""),
+                    sql.SQL(" CASCADE" if cascade else "")
+                )
 
-            # Execute the delete operation
-            cur.execute(delete_query)
-            logging.info(f"All features in table {schema}.{table_name} deleted successfully.")
-        
-        except Exception as delete_error:
-            logging.error(f"Failed to delete features in table {schema}.{table_name}. Error: {str(delete_error)}")
-        
-    finally:
-        # Clean up: close the cursor and the connection
-        cur.close()
-        conn.close()
+                # Execute the truncate operation
+                cur.execute(query)
+                logging.info(
+                    "Table %s.%s truncated successfully.", schema, table_name)
+        except psycopg2.errors.FeatureNotSupported as e:
+            if "cannot truncate a table referenced in a foreign key constraint" in str(e):
+                # If an error occurs during truncation, log the error and attempt to delete instead
+                logging.warning(
+                    """Failed to truncate table %s.%s, perhaps because of a dependency within the database. 
+                    Trying to delete features without truncating. Error: %s""", schema, table_name, str(e))
+
+                try:
+                    with conn.cursor() as cur:
+                        # Construct the DELETE SQL query
+                        delete_query = sql.SQL("DELETE FROM {}.{}").format(
+                            sql.Identifier(schema),
+                            sql.Identifier(table_name)
+                        )
+
+                        # Execute the delete operation
+                        cur.execute(delete_query)
+                        logging.info(
+                            "All features in table %s.%s deleted successfully.", schema, table_name)
+
+                        if reset_sequence:
+                            logging.warning(
+                                "Resetting sequence after deletion is not currently supported.")
+                except Exception as delete_error:
+                    logging.error(
+                        "Failed to delete features in table %s.%s. Error: %s",
+                        schema, table_name, str(delete_error))
+                    raise delete_error
+            else:
+                raise e
+        except Exception as e:
+            logging.error(
+                "Failed to truncate features in table %s.%s. Error: %s", schema, table_name, str(e))
+            raise e
